small fixes
I want to get this all commited because I won't come back to it for like 6-8 hours and by then I'll be tired of committing and just not even when I work on it. I have other projects and work to do and it's almost 2:30 already so I should go work on those and stop screwing around with this scouting app. I bet no one in the world reads these and I find it only marginally more likely that this comment is even included in the commit somewhere online. If someone is reading, hi! Anyways I'm gonna go work on something not invovling networking but definitely involving Rust on my 3am insomnia and depression fueled coding spree where I'm more prodcutive then I've been in months.  Trans Rights. Henry fix the stupid errors gradle is giving on Robot2020. I swear I'll get it connected to the app if you just fix those dumb errors. If you need to switch to windows or disable unix whatever just get it done. I'm claiming I don't know why my phone doesn't work as well these days but I definitely splilled water in it. This seems like a waste of time but supprisingly it's only been 3 minuites to type all of this. Oh, and the 2020 game looks complicated and dumb so yeah. Is this the kind of thing future employers will see and not hire me for? Doubt it, if it is totally worth it. And here I'm gonna commit this and find out these comments get emailed to Jason or something and I'm gonna get in trouble. Who would even be mad over this though? Epstine didn't kill himself.