DWRNT
Do We Really Need This: everything edition

A ton of the code was repetitive and unneeded. As this is pretty much a guess and check project I had some better code the second time around trying to write the website with some more experience. Yes, I know this rewrote most if not all of the code. The network code can still be improved and compressed, this is for the future. For now I'm going to work on completion of the project because truly I cannot work on network code at 4AM for gods sake

I wish you all the best,
   -Jaiden